refactor: create folder for Dashboard
